Fix initialization of worker.
The first argument to this app's App() constructors is used as second
argument for the Flask() constructor, thus trying to set the
`static_url_path` and not the `import_name`.
There is no need to set the `import_name` in `worker.py`, this is done
hardcoded in the code already.